+++ src/plugins/data_source/public/types.ts
@@ -3,8 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface DataSourcePluginSetup {}
+export interface DataSourcePluginSetup {
+  dataSourceEnabled: boolean;
+  hideLocalCluster: boolean;
+}
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface DataSourcePluginStart {}
+export interface DataSourcePluginStart {
+  dataSourceEnabled: boolean;
+  hideLocalCluster: boolean;
+}