+++ src/plugins/dev_tools/public/application.tsx
@@ -56,6 +56,7 @@ interface DevToolsWrapperProps {
   savedObjects: SavedObjectsStart;
   notifications: NotificationsStart;
   dataSourceEnabled: boolean;
+  hideLocalCluster: boolean;
 }
 
 interface MountedDevToolDescriptor {
@@ -71,6 +72,7 @@ function DevToolsWrapper({
   savedObjects,
   notifications: { toasts },
   dataSourceEnabled,
+  hideLocalCluster,
 }: DevToolsWrapperProps) {
   const mountedTool = useRef<MountedDevToolDescriptor | null>(null);
 
@@ -136,6 +138,7 @@ function DevToolsWrapper({
               notifications={toasts}
               onSelectedDataSource={onChange}
               disabled={!dataSourceEnabled}
+              hideLocalCluster={hideLocalCluster}
               fullWidth={false}
             />
           </div>
@@ -153,7 +156,12 @@ function DevToolsWrapper({
               mountedTool.current.devTool !== activeDevTool ||
               mountedTool.current.mountpoint !== element)
           ) {
-            await remount(element);
+            let initialDataSourceId;
+            if (!dataSourceEnabled || (dataSourceEnabled && !hideLocalCluster)) {
+              initialDataSourceId = '';
+            }
+
+            await remount(element, initialDataSourceId);
           }
         }}
       />
@@ -212,6 +220,7 @@ export function renderApp(
   { dataSource }: DevToolsSetupDependencies
 ) {
   const dataSourceEnabled = !!dataSource;
+  const hideLocalCluster = dataSource?.hideLocalCluster ?? false;
   if (redirectOnMissingCapabilities(application)) {
     return () => {};
   }
@@ -241,6 +250,7 @@ export function renderApp(
                     savedObjects={savedObjects}
                     notifications={notifications}
                     dataSourceEnabled={dataSourceEnabled}
+                    hideLocalCluster={hideLocalCluster}
                   />
                 )}
               />